protocol conformances & enum conversions
Everything now conforms to Equatable and Hashable (where applicable)

Also I'm not happy with the structs with static members conversion I did in
a previous commit, so I reverted that. Potentially adding new members might
theoretically be a breaking api change for now, but my fingers are crossed
for a Swift proposal that might be able to fix that in the close future.